% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{convert_magrittr_into_topic}
\alias{convert_magrittr_into_topic}
\title{Obtain correct docs word for magrittr pipes}
\usage{
convert_magrittr_into_topic(pipe)
}
\description{
Obtain correct docs word for magrittr pipes
}
\keyword{internal}
