% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\name{convert_operator_into_topic}
\alias{convert_operator_into_topic}
\title{Obtain correct category and topic for any operator get_help() arguments}
\usage{
convert_operator_into_topic(operator)
}
\description{
Obtain correct category and topic for any operator get_help() arguments
}
\keyword{internal}
