% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{convert_topic_into_magrittr}
\alias{convert_topic_into_magrittr}
\title{Obtain correct operator for topic pipe words}
\usage{
convert_topic_into_magrittr(topic)
}
\description{
Obtain correct operator for topic pipe words
}
\keyword{internal}
