% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Crayon cat operator}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{A \code{crayon} value}

\item{rhs}{A \code{crayon} value}
}
\value{
The combined \code{crayon} values
}
\description{
Crayon cat operator
}
\keyword{internal}
