% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{dplyr_topics}
\alias{dplyr_topics}
\title{All the dplyr topics}
\format{
An object of class \code{character} of length 18.
}
\usage{
dplyr_topics
}
\description{
All the dplyr topics
}
\keyword{internal}
