% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{find_topic_category}
\alias{find_topic_category}
\title{Function to identify a topic's category}
\usage{
find_topic_category(topic)

find_topic_category(topic)
}
\arguments{
\item{topic}{The topic of interest}
}
\value{
The category

The category
}
\description{
Function to identify a topic's category

Function to identify a topic's category
}
