% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_second_tibble.R
\docType{data}
\name{first_tibble}
\alias{first_tibble}
\title{A small tibble used to demonstrate the \code{{dplyr}} join functions, derived from \code{msleep}}
\format{
A tibble (data frame) with 4 rows and 3 variables:
\describe{
\item{name}{common name of the carnivore species}
\item{vore}{what it eats}
\item{conservation}{conservation status}
}
}
\source{
\url{https://ggplot2.tidyverse.org/}
}
\usage{
first_tibble
}
\description{
A dataset containing name, vore, and conservation for 4 mammal species, obtained from \code{msleep}
}
\examples{
\dontrun{
  first_tibble
}
}
\keyword{datasets}
