% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{forcats_topics}
\alias{forcats_topics}
\title{All the forcats topics}
\format{
An object of class \code{character} of length 7.
}
\usage{
forcats_topics
}
\description{
All the forcats topics
}
\keyword{internal}
