% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{html_topics_path}
\alias{html_topics_path}
\title{Internal path to html topics}
\format{
An object of class \code{character} of length 1.
}
\usage{
html_topics_path
}
\description{
Internal path to html topics
}
\keyword{internal}
