% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{introverse_viewer}
\alias{introverse_viewer}
\title{Viewer wrapper for showing the help docs}
\usage{
introverse_viewer(html_file)
}
\arguments{
\item{html_file}{The (temporary) HTML file containing the rendered help to view}
}
\description{
Viewer wrapper for showing the help docs
}
\keyword{internal}
