% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{launch_interactive_help}
\alias{launch_interactive_help}
\title{Launch the learnr interactive help topic HTML}
\usage{
launch_interactive_help(topic, category)
}
\arguments{
\item{topic}{The topic to launch}

\item{category}{The topic to launch's category}
}
\description{
Launch the learnr interactive help topic HTML
}
