% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{launch_standalone_help}
\alias{launch_standalone_help}
\title{Launch (show) the standalone help topic HTML content in the viewer pane}
\usage{
launch_standalone_help(topic, category)
}
\arguments{
\item{topic}{The topic to launch}

\item{category}{The topic to launch's category}
}
\description{
Launch (show) the standalone help topic HTML content in the viewer pane
}
