% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{learnr_chunk_string}
\alias{learnr_chunk_string}
\title{String to add to an Rmd file to include the learnr library}
\format{
An object of class \code{character} of length 1.
}
\usage{
learnr_chunk_string
}
\description{
String to add to an Rmd file to include the learnr library
}
\keyword{internal}
