% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{learnr_topics_path}
\alias{learnr_topics_path}
\title{Internal path to Rmd topics formatted as learnr.}
\format{
An object of class \code{character} of length 1.
}
\usage{
learnr_topics_path
}
\description{
Internal path to Rmd topics formatted as learnr.
}
\keyword{internal}
