% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{magrittr_operators}
\alias{magrittr_operators}
\title{Magrittr operators}
\format{
An object of class \code{character} of length 2.
}
\usage{
magrittr_operators
}
\description{
Magrittr operators
}
\keyword{internal}
