% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{magrittr_topics}
\alias{magrittr_topics}
\title{Magrittr topics}
\format{
An object of class \code{character} of length 2.
}
\usage{
magrittr_topics
}
\description{
Magrittr topics
}
\keyword{internal}
