% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{message_show_categories}
\alias{message_show_categories}
\title{Redirect user to the \code{show_topics()} output if they entered a category into \code{get_help()}}
\usage{
message_show_categories(category)
}
\arguments{
\item{category}{The category to whose topics this function reveals}
}
\description{
Redirect user to the \code{show_topics()} output if they entered a category into \code{get_help()}
}
\keyword{internal}
