% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msleep.R
\docType{data}
\name{msleep}
\alias{msleep}
\title{Dataset: An updated and expanded version of the mammals sleep dataset, exported from \code{ggplot2}}
\format{
A tibble (data frame) with 83 rows and 11 variables:
\describe{
\item{name}{common name}
\item{genus}{the Genus classification of the species}
\item{vore}{carnivore, omnivore, herbivore, or insectivore?}
\item{order}{the Order classification of the species}
\item{conservation}{the conservation status of the species}
\item{sleep_total}{total amount of sleep, in hours}
\item{sleep_rem}{rem sleep, in hours}
\item{sleep_cycle}{length of sleep cycle, in hours}
\item{awake}{amount of time spent awake, in hours}
\item{brainwt}{brain weight in kilograms}
\item{bodywt}{body weight in kilograms}
}
}
\source{
\url{https://ggplot2.tidyverse.org/}
}
\usage{
msleep
}
\description{
This is the msleep dataset exported from the \code{ggplot2} package version 3.3.5. For more information, see \code{?ggplot2::msleep}.
}
\keyword{datasets}
