% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{operators}
\alias{operators}
\title{Operators documented in logical, math, or assignment help pages}
\format{
An object of class \code{character} of length 18.
}
\usage{
operators
}
\description{
Operators documented in logical, math, or assignment help pages
}
\keyword{internal}
