% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get-help.R
\name{quick_help}
\alias{quick_help}
\title{Get quick help for a given function}
\usage{
quick_help(f)
}
\arguments{
\item{f}{A function whose introverse quick docs to look up}
}
\description{
Get quick help for a given function
}
