% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{readr_topics}
\alias{readr_topics}
\title{All the readr topics}
\format{
An object of class \code{character} of length 8.
}
\usage{
readr_topics
}
\description{
All the readr topics
}
\keyword{internal}
