% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{redirect_topic}
\alias{redirect_topic}
\title{For topics with shared pages, redirect a given topic to the name of the topic that has an Rmd page}
\usage{
redirect_topic(topic)
}
\arguments{
\item{topic}{The topic of interest}
}
\value{
Either the provided topic if it is not shared, or the topic whose page exists if it is shared
}
\description{
For topics with shared pages, redirect a given topic to the name of the topic that has an Rmd page
}
\keyword{internal}
