% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-render_rmd.R
\name{render_single_rmd_topic_to_learnr}
\alias{render_single_rmd_topic_to_learnr}
\title{Render an Rmd help topic page into a learnr tutorial HTML}
\usage{
render_single_rmd_topic_to_learnr(rmd_file, learnr_file, output_path)
}
\description{
Render an Rmd help topic page into a learnr tutorial HTML
}
\keyword{internal}
