% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-render_rmd.R
\name{render_single_rmd_topic_to_standalone}
\alias{render_single_rmd_topic_to_standalone}
\title{Render an Rmd help topic page into a standalone HTML}
\usage{
render_single_rmd_topic_to_standalone(rmd_file, html_file)
}
\description{
Render an Rmd help topic page into a standalone HTML
}
\keyword{internal}
