% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{reveal_help}
\alias{reveal_help}
\title{Reveal the help topic HTML content in the viewer pane}
\usage{
reveal_help(topic)
}
\arguments{
\item{topic}{The topic to reveal}
}
\description{
Reveal the help topic HTML content in the viewer pane
}
