% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{rmd_topics_path}
\alias{rmd_topics_path}
\title{Internal path to Rmd topics}
\format{
An object of class \code{character} of length 1.
}
\usage{
rmd_topics_path
}
\description{
Internal path to Rmd topics
}
\keyword{internal}
