% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_second_tibble.R
\docType{data}
\name{second_tibble}
\alias{second_tibble}
\title{Small tibble used to demonstrate joins}
\format{
A tibble (data frame) with 4 rows and 3 variables:
\describe{
\item{name}{common name of the carnivore species}
\item{vore}{what it eats}
\item{order}{taxonomic order}
}
}
\source{
\url{https://ggplot2.tidyverse.org/}
}
\usage{
second_tibble
}
\description{
A dataset containing name, vore, and order for 4 mammal species, obtained from \code{msleep}
}
\keyword{datasets}
