% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{shared_doc_pages}
\alias{shared_doc_pages}
\title{List of which topics share a given page, for redirection. The FIRST one represents the rmd file which \emph{exist}}
\format{
An object of class \code{list} of length 21.
}
\usage{
shared_doc_pages
}
\description{
List of which topics share a given page, for redirection. The FIRST one represents the rmd file which \emph{exist}
}
\keyword{internal}
