% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{show_topic_categories}
\alias{show_topic_categories}
\title{Show the list of topic categories}
\usage{
show_topic_categories()
}
\value{
Prints out available topic categories
}
\description{
TODO: Should we reveal this in Console or Viewer?
}
\examples{
\dontrun{
show_topic_categories()
}
}
