% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\name{show_topics}
\alias{show_topics}
\title{Show the topic list}
\usage{
show_topics(category = NULL)
}
\arguments{
\item{category}{If specified, gives only topics in the given category (name in topic_list)}
}
\value{
Prints out available topics by package
}
\description{
Show the topic list
}
\examples{
\dontrun{
show_topics() # Show a list of topics, organized by category
# Show all topics in the `dplyr` category
show_topics(category = 'dplyr') 
# Show all topics in the `dplyr` and `tidyr` categories
show_topics(category = c('dplyr', 'tidyr')) 
}
}
