% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{stringr_topics}
\alias{stringr_topics}
\title{All the stringr topics}
\format{
An object of class \code{character} of length 10.
}
\usage{
stringr_topics
}
\description{
All the stringr topics
}
\keyword{internal}
