% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get-help.R
\name{style_comment}
\alias{style_comment}
\title{Style text as code comment}
\usage{
style_comment(...)
}
\arguments{
\item{x}{Text to be styled as a code comment}
}
\description{
Style text as code comment
}
