% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{tidyr_topics}
\alias{tidyr_topics}
\title{All the tidyr topics}
\format{
An object of class \code{character} of length 5.
}
\usage{
tidyr_topics
}
\description{
All the tidyr topics
}
\keyword{internal}
