% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-topics.R
\docType{data}
\name{topic_list}
\alias{topic_list}
\title{List of available help topics by category}
\format{
An object of class \code{list} of length 1.

An object of class \code{list} of length 1.
}
\usage{
topic_list

topic_list
}
\description{
List of available help topics by category

List of available help topics by category
}
\keyword{datasets}
