% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{topic_width}
\alias{topic_width}
\title{Internal options width for HTML size}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
topic_width
}
\description{
Internal options width for HTML size
}
\keyword{internal}
