% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{base_topics}
\alias{base_topics}
\title{All the base R topics}
\format{
An object of class \code{character} of length 34.
}
\usage{
base_topics
}
\description{
All the base R topics
}
\keyword{internal}
