% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{get_help}
\alias{get_help}
\title{Prints help and examples for a given topic to the R Console}
\usage{
get_help(name_of_topic = NULL)
}
\arguments{
\item{name_of_topic}{A topic whose introverse quick docs to look up}
}
\description{
Prints help and examples for a given topic to the R Console
}
