% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get-help.R
\name{get_help}
\alias{get_help}
\title{Prints help and examples for a given function to the R Console}
\usage{
get_help(f = NULL)
}
\arguments{
\item{f}{A function whose introverse quick docs to look up}
}
\description{
Prints help and examples for a given function to the R Console
}
