% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{get_help}
\alias{get_help}
\title{Use this function to get help on a topic of interest}
\usage{
get_help(topic = NULL, browser = FALSE)
}
\arguments{
\item{topic}{A topic whose introverse quick docs to look up}

\item{browser}{Whether to display in browser. By default, when in an RStudio session, it is displayed in the Viewer pane}
}
\description{
Use this function to get help on a topic of interest
}
