% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-topics.R
\docType{data}
\name{ggplot2_topics}
\alias{ggplot2_topics}
\title{All the ggplot2 topics}
\format{
An object of class \code{character} of length 10.
}
\usage{
ggplot2_topics
}
\description{
All the ggplot2 topics
}
\keyword{internal}
