% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-get_help.R
\name{reveal_help}
\alias{reveal_help}
\title{Reveal the standalone help topic HTML content in the viewer pane}
\usage{
reveal_help(category, topic)
}
\arguments{
\item{category}{The topic to launch's category}

\item{topic}{The topic to launch}
}
\description{
Reveal the standalone help topic HTML content in the viewer pane
}
